Inserting process intance into logs after each process request
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C57766

+++ SOM/Code/BPMExtended.Main.Business/Billing/CustomerRequestManager.cs
@@ -391,7 +391,7 @@ namespace BPMExtended.Main.Business
             {
                 var processInstanceLog = new SOMProcessInstanceLog()
                 {
-                    ProcessInstanceId = (long)entities[i].GetColumnValue("StProcessInstanceId"),
+                    ProcessInstanceId = (int)entities[i].GetColumnValue("StProcessInstanceId"),
                     RequestId = requestId
                 };
                 processInstancesLogs.Add(processInstanceLog);
